requirements.txt
Django==2.2.5
django-tinymce4-lite==1.7.5
jsmin==2.2.2
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
